requirements.txt
absl-py==1.3.0
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bert-tensorflow==1.0.1
bleach==5.0.1
blessed==1.19.1
blinker==1.4
Brlapi==0.8.3
CacheControl==0.12.11
cachetools==5.2.0
cachy==0.3.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
cleo==0.8.1
click==8.0.3
clikit==0.6.2
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.5
crashtest==0.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
entrypoints==0.4
executing==1.1.1
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.6.0
flatbuffers==22.10.26
fonttools==4.37.4
future==0.18.2
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==1.0.0
grpcio==1.50.0
h5py==3.7.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras==2.11.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.4
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nbresult==0.0.8
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.4.12
numpy==1.23.4
nvidia-ml-py==11.495.46
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkginfo==1.8.2
platformdirs==2.5.1
poetry==1.1.12
poetry-core==1.0.7
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.12.4
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pylev==1.2.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
SecretStorage==3.3.1
Send2Trash==1.8.0
shellingham==1.4.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
stack-data==0.5.1
systemd-python==234
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
terminado==0.16.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tomlkit==0.9.2
tornado==6.2
traitlets==5.4.0
typing_extensions==4.4.0
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xdg==5
xgboost==1.7.1
xkit==0.0.0
zipp==1.0.0
